full changes

+++ Задача_6/Program.cs
@@ -1,5 +1,5 @@
 ﻿Console.WriteLine("Введите число:");
-double number = Convert.ToDouble(Console.ReadLine());
+int number = Convert.ToInt32(Console.ReadLine());
 if (number % 2 == 0)
 {
     Console.WriteLine("Да");
